requirements.txt
# Orchestrator Requirements

# Supabase
supabase>=2.0.0
python-dotenv>=1.0.0

# Google Drive API (Optional)
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.100.0
